requirements.txt
accelerate==1.2.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
asttokens==3.0.0
attrs==24.2.0
azure-core==1.32.0
azure-identity==1.19.0
azure-storage-blob==12.24.0
azure-storage-file-datalake==12.18.0
bcrypt==4.2.1
blinker==1.9.0
boto3==1.35.79
botocore==1.35.79
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
circuitbreaker==2.0.0
click==8.1.7
cloudpickle==3.1.0
contourpy==1.3.1
cramjam==2.9.0
cryptography==44.0.0
cycler==0.12.1
databricks-sdk==0.38.0
datasets==3.1.0
decorator==5.1.1
Deprecated==1.2.15
dill==0.3.8
docker==7.1.0
docker-pycreds==0.4.0
einops==0.8.0
executing==2.1.0
filelock==3.16.1
flash-attn==2.7.2.post1
Flask==3.1.0
fonttools==4.55.2
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.0
google-auth==2.36.0
google-cloud-core==2.4.1
google-cloud-storage==2.10.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
graphene==3.4.3
graphql-core==3.2.5
graphql-relay==3.2.0
greenlet==3.1.1
gunicorn==23.0.0
huggingface-hub==0.26.5
idna==3.10
importlib_metadata==8.5.0
ipython==8.30.0
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.7
lightning==2.4.0
lightning-utilities==0.11.9
Mako==1.3.8
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
mlflow==2.18.0
mlflow-skinny==2.18.0
mosaicml-streaming==0.10.0
mpmath==1.3.0
msal==1.31.1
msal-extensions==1.2.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.4.2
ninja==1.11.1.2
numpy==2.2.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oci==2.140.0
opentelemetry-api==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
packaging==24.2
pandas==2.2.3
paramiko==3.5.0
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
portalocker==2.10.1
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.1
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==24.3.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-snappy==0.7.3
pytorch-lightning==2.4.0
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
s3transfer==0.10.4
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentry-sdk==2.19.2
setproctitle==1.3.4
setuptools==75.6.0
six==1.17.0
smmap==5.0.1
SQLAlchemy==2.0.36
sqlparse==0.5.2
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1
torchmetrics==1.6.0
torchvision==0.20.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.0
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wandb==0.19.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.0
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstd==1.5.5.1
